% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{account_player}
\alias{account_player}
\title{Account Metadata}
\usage{
account_player(account_id)
}
\arguments{
\item{account_id}{A unique identifier for an individual's account.}
}
\description{
Provides metadata for an individual's account including a list of profiles associated
with the account, as well as their their respective display names, profile URLs
and avatar URLs.
}
