% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{authentication}
\alias{authentication}
\alias{get_token}
\alias{valid_token}
\alias{set_token}
\alias{remove_token}
\title{Authentication Methods}
\usage{
get_token(client_id, client_secret)

valid_token(access_token)

set_token(access_token)

remove_token()
}
\arguments{
\item{client_id, client_secret}{A client ID and client secret can be obtained from the
\href{https://develop.battle.net/}{Blizzard Developer Portal}. For more information on
creating a client, visit
\href{https://develop.battle.net/documentation/guides/getting-started}{Getting Started}.}

\item{access_token}{An OAuth 2.0 access token required to use the Blizzard API. Access tokens can be
obtained by using \code{get_token()} with a valid client ID and client secret.}
}
\description{
The Blizzard API uses OAuth 2.0 for authorization. For more information on how Blizzard
   uses OAuth in their API, visit \url{https://develop.battle.net/documentation/guides/using-oauth}.

   Before using the Blizzard API, one must first create a client in the
   \href{https://develop.battle.net/}{Blizzard Developer Portal} and obtain a valid client ID and
   client secret. For more information on getting started, see:
   \href{https://develop.battle.net/documentation/guides/getting-started}{Getting Started}.

   Once a client has been created, use \code{get_token(client_id,client_secret)}
   to obtain a valid access token. This token can then be supplied to
   \code{set_token()} to set an environment variable
   for all future API calls.

   Once \code{set_token()} has been used, an access token can be removed from the environment
   using \code{remove_token()}

   Note that access tokens are set to expire in 24 hours and, subsequently,
   a new token must be used for any future API calls.
}
\note{
Access tokens expire after 24 hours.
}
