% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{blizzard}
\alias{blizzard}
\title{Blizzard API Calls}
\description{
A list of functions to make calls to the Blizzard API. Before using any of the
   functions, ensure that an access token is set as an environment variable by using
   \code{\link{set_token}}.

   Explanations for each of the functions can be found in the
   \href{https://develop.battle.net/documentation/starcraft-2}{Starcraft II API documentation}.

   Note that the China region is \strong{not} supported at this time.
}
