% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{get_player}
\alias{get_player}
\title{Account Metadata}
\usage{
get_player(account_id, host_region = "us")
}
\arguments{
\item{account_id}{A unique identifier for an individual's account.}

\item{host_region}{The host region that the API call will be sent to. For most API calls, the same data will be
returned regardless of which region the request is sent to. Must be one of "us", "eu", "kr", "tw", "cn". For more
information on regionality, refer to
\href{https://develop.battle.net/documentation/guides/regionality-and-apis}{Regionality and APIs}.}
}
\description{
Provides metadata for an individual's account including a list of profiles associated
with the account, as well as their their respective display names, profile URLs
and avatar URLs.
}
\references{
\itemize{
   \item \href{https://develop.battle.net/documentation/starcraft-2/community-apis}{Blizzard Community API Documentation}
   \item \href{https://develop.battle.net/documentation/guides/regionality-and-apis}{Regionality and APIs}
   }
}
