% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{ladder_data}
\alias{ladder_data}
\title{Ladder Data}
\usage{
ladder_data(ladder_id)
}
\arguments{
\item{ladder_id}{A unique identifier for a particular ladder. With the exception of gradmaster, leagues
(bronze, silver, etc.), are separated into tiers (1,2,3) which are further separated into divisions.
These divisions, or ladders, each have a unique identifier.}
}
\description{
Provides data of players in a particular ladder. This includes MMR, points,
win/loss record, time of joining, time of a player's last game, and more.
}
