% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{ladder_season}
\alias{ladder_season}
\title{Current Season Information}
\usage{
ladder_season(region_id)
}
\arguments{
\item{region_id}{A numeric argument indiciating in which region the profile resides.
    \itemize{
        \item 1 = US Region
        \item 2 = EU Region
        \item 3 = KR/TW Region
}}
}
\description{
Provides the current season ID, starting date, and ending date.
}
\seealso{
Other ladder data API calls: 
\code{\link{ladder_gm_leaderboard}()}
}
\concept{ladder data API calls}
