% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{league_get_league_data}
\alias{league_get_league_data}
\title{League Data}
\usage{
league_get_league_data(season_id, queue_id, team_type, league_id)
}
\arguments{
\item{season_id}{A numeric argument indicating a particular ladder season. The Blizzard API
only accepts season 28 and higher (i.e. data prior to this season is unavailable).}

\item{queue_id}{\itemize{
   \item 1 = WoL 1v1
   \item 2 = WoL 2v2
   \item 3 = WoL 3v3
   \item 4 = WoL 4V4
   \item 101 = HotS 1v1
   \item 102 = HotS 2v2
   \item 103 = HotS 3v3
   \item 104 = HotS 4v4
   \item 201 = LotV 1v1
   \item 202 = LotV 2v2
   \item 203 = LotV 3v3
   \item 204 = LotV 4v4
   \item 206 = LotV Archon
}}

\item{team_type}{\itemize{
    \item 0 = Arranged
    \item 1 = Random
}}

\item{league_id}{\itemize{
   \item 0 = Bronze
   \item 1 = Silver
   \item 2 = Gold
   \item 3 = Platinum
   \item 4 = Diamond
   \item 5 = Masters
   \item 6 = Grandmaster
}}
}
\description{
League data is divided into 3 tiers for each league (with the exception of grandmaster, which only has
1 tier) and further divided into a number of divisions depending on how many players are in a given league.
League data contains the number of divisions, the unique ladder ID of each division and the
total player count contained within each division.
}
\note{
League data is only available for season 28 and higher.
}
