% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{legacy_achievements}
\alias{legacy_achievements}
\title{Legacy API - Available Achievements}
\usage{
legacy_achievements(region_id)
}
\arguments{
\item{region_id}{A numeric argument indiciating in which region the profile resides.
    \itemize{
        \item 1 = US Region
        \item 2 = EU Region
        \item 3 = KR/TW Region
}}
}
\description{
Provides a listing of available achievements in Starcraft II.
}
\seealso{
\code{\link{profile_static}}

Other legacy API calls: 
\code{\link{legacy_ladders}()},
\code{\link{legacy_ladder}()},
\code{\link{legacy_match_history}()},
\code{\link{legacy_profile}()},
\code{\link{legacy_rewards}()}
}
\concept{legacy API calls}
