% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{legacy_ladder}
\alias{legacy_ladder}
\title{Legacy API - Ladder Details}
\usage{
legacy_ladder(region_id, ladder_id)
}
\arguments{
\item{region_id}{A numeric argument indiciating in which region the profile resides.
    \itemize{
        \item 1 = US Region
        \item 2 = EU Region
        \item 3 = KR/TW Region
}}

\item{ladder_id}{A unique identifier for a particular ladder. With the exception of gradmaster, leagues
(bronze, silver, etc.), are separated into tiers (1,2,3) which are further separated into divisions.
These divisions, or ladders, each have a unique identifier.}
}
\description{
Provides a listing of players in a given ladder. Also provides other information such as
their ladder record, points, profile information, and clan.
}
\seealso{
Other legacy API calls: 
\code{\link{legacy_achievements}()},
\code{\link{legacy_ladders}()},
\code{\link{legacy_match_history}()},
\code{\link{legacy_profile}()},
\code{\link{legacy_rewards}()}
}
\concept{legacy API calls}
