% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{legacy_ladders}
\alias{legacy_ladders}
\title{Legacy API - Profile Ladder Summary}
\usage{
legacy_ladders(region_id, realm_id, profile_id)
}
\arguments{
\item{region_id}{A numeric argument indiciating in which region the profile resides.
    \itemize{
        \item 1 = US Region
        \item 2 = EU Region
        \item 3 = KR/TW Region
}}

\item{realm_id}{A numeric argument indiciating in which realm the profile resides. A realm is a subset
of the region.
 \itemize{
     \item{US Region}
         \itemize{
             \item{1 = US}
             \item{2 = LatAm}
         }
     \item{EU Region}
         \itemize{
             \item{1 = Europe}
             \item{2 = Russia}
     }
     \item{KR/TW Region}
         \itemize{
             \item{1 = Korea}
             \item{2 = Taiwan}
     }
 }}

\item{profile_id}{A unique, numeric identifier for an individual's profile.}
}
\description{
Provides information about a profile's performance in the current season, previous season, and
showcase entries.
}
\seealso{
\code{\link{profile_ladder_summary}}

Other legacy API calls: 
\code{\link{legacy_achievements}()},
\code{\link{legacy_ladder}()},
\code{\link{legacy_match_history}()},
\code{\link{legacy_profile}()},
\code{\link{legacy_rewards}()}
}
\concept{legacy API calls}
