% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{profile_ladder}
\alias{profile_ladder}
\title{Ladder Details and Profile Rank}
\usage{
profile_ladder(region_id, realm_id, profile_id, ladder_id)
}
\arguments{
\item{region_id}{A numeric argument indiciating in which region the profile resides.
    \itemize{
        \item 1 = US Region
        \item 2 = EU Region
        \item 3 = KR/TW Region
}}

\item{realm_id}{A numeric argument indiciating in which realm the profile resides. A realm is a subset
of the region.
 \itemize{
     \item{US Region}
         \itemize{
             \item{1 = US}
             \item{2 = LatAm}
         }
     \item{EU Region}
         \itemize{
             \item{1 = Europe}
             \item{2 = Russia}
     }
     \item{KR/TW Region}
         \itemize{
             \item{1 = Korea}
             \item{2 = Taiwan}
     }
 }}

\item{profile_id}{A unique, numeric identifier for an individual's profile.}

\item{ladder_id}{A unique identifier for a particular ladder. With the exception of gradmaster, leagues
(bronze, silver, etc.), are separated into tiers (1,2,3) which are further separated into divisions.
These divisions, or ladders, each have a unique identifier.}
}
\description{
Provides information about a particular ladder and the individual's rank and status within that
ladder (i.e. rank, MMR, etc.).
}
\seealso{
Other profile API calls: 
\code{\link{profile_ladder_summary}()},
\code{\link{profile_metadata}()},
\code{\link{profile_profile}()},
\code{\link{profile_static}()}
}
\concept{profile API calls}
