% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{profile_metadata}
\alias{profile_metadata}
\title{Profile Metadata}
\usage{
profile_metadata(region_id, realm_id, profile_id)
}
\arguments{
\item{region_id}{A numeric argument indiciating in which region the profile resides.
    \itemize{
        \item 1 = US Region
        \item 2 = EU Region
        \item 3 = KR/TW Region
}}

\item{realm_id}{A numeric argument indiciating in which realm the profile resides. A realm is a subset
of the region.
 \itemize{
     \item{US Region}
         \itemize{
             \item{1 = US}
             \item{2 = LatAm}
         }
     \item{EU Region}
         \itemize{
             \item{1 = Europe}
             \item{2 = Russia}
     }
     \item{KR/TW Region}
         \itemize{
             \item{1 = Korea}
             \item{2 = Taiwan}
     }
 }}

\item{profile_id}{A unique, numeric identifier for an individual's profile.}
}
\description{
Provides metadata for an individual's profile including their display name, profile URL,
and avatar URL.
}
\seealso{
Other profile API calls: 
\code{\link{profile_ladder_summary}()},
\code{\link{profile_ladder}()},
\code{\link{profile_profile}()},
\code{\link{profile_static}()}
}
\concept{profile API calls}
