% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{profile_static}
\alias{profile_static}
\title{Static Profile Data}
\usage{
profile_static(region_id)
}
\arguments{
\item{region_id}{A numeric argument indiciating in which region the profile resides.
    \itemize{
        \item 1 = US Region
        \item 2 = EU Region
        \item 3 = KR/TW Region
}}
}
\description{
Provides static information (achievements, categories, criteria, and rewards) about SC2 profiles
in a given region.
}
\seealso{
Other profile API calls: 
\code{\link{profile_ladder_summary}()},
\code{\link{profile_ladder}()},
\code{\link{profile_metadata}()},
\code{\link{profile_profile}()}
}
\concept{profile API calls}
