% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc2api.R
\docType{package}
\name{SC2API}
\alias{SC2API}
\title{\code{SC2API} package}
\description{
A simple wrapper written in R for Blizzard's Starcraft II API. Enables users to retrieve various data pertaining to
accounts, profiles, seasons and ladders.
}
\details{
Before using, it is recommended to view the README on
\href{https://github.com/SamMorrissette/sc2api#readme}{Github}

Further documentation can be found in Blizzard's
\href{https://develop.battle.net/documentation/starcraft-2}{Starcraft II API documentation}.
}
